Weeks1-tasks
Project Description: The Success Squad

"The Success Squad" is an online course website built using HTML, CSS, and JavaScript. The platform offers users access to a variety of educational courses aimed at personal and professional growth. The design focuses on providing a user-friendly experience, with responsive layouts, interactive features, and seamless navigation. 

Key features of the site include:
- A clean and modern interface created with HTML and CSS for a professional look and feel.
- Interactive elements and dynamic content powered by JavaScript to enhance user engagement.
- A responsive design that adapts to different devices and screen sizes, ensuring accessibility for all users.

This project aims to deliver a high-quality e-learning platform that empowers individuals to achieve success through knowledge and skill development.
This is only the Frontend part of the website.